requirements.txt
# Core Web Framework (using Flask instead of FastAPI for Python 3.13 compatibility)
flask==3.0.0
flask-cors==4.0.0

# Core Utilities
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
httpx==0.23.3

# Document Processing
beautifulsoup4==4.12.2
html2text==2020.1.16
pypdf2==3.0.1
python-docx==1.1.0

# OpenAI API
openai==0.27.8

# Development Tools
black==23.12.1
isort==5.13.2 